// packages/datadog-api-client-v1/models/UsageBillableSummaryBody.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Response with properties for each aggregated usage type.
*/
export declare class UsageBillableSummaryBody {
    /**
    * The total account usage.
    */
    'accountBillableUsage'?: number;
    /**
    * Elapsed usage hours for some billable product.
    */
    'elapsedUsageHours'?: number;
    /**
    * The first billable hour for the org.
    */
    'firstBillableUsageHour'?: Date;
    /**
    * The last billable hour for the org.
    */
    'lastBillableUsageHour'?: Date;
    /**
    * The number of units used within the billable timeframe.
    */
    'orgBillableUsage'?: number;
    /**
    * The percentage of account usage the org represents.
    */
    'percentageInAccount'?: number;
    /**
    * Units pertaining to the usage.
    */
    'usageUnit'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): UsageBillableSummaryBody;
    static serialize(data: UsageBillableSummaryBody): {
        [key: string]: any;
    };
    constructor();
}